requirements.txt
requests
Flask
psycopg2-binary
python-dotenv
celery
redis